// src/interfaces/AppInterfaces/KeyManBeatInterface/KeymanBeatDataTableInterface.ts
export interface IKeymanBeatDataTableInterface {
  sectionName: string;
  beatId: string;
  activeStatus: boolean;
  startTime: number;
  endTime: number;
  deviceImei: number;
  deviceTypeId: number;
  shiftType: number;
  tendKm: number;
  bendTime: number;
  tstartKm: number;
  bstartTime: number;
  deviceNo: number;
  deviceName: string;
}